Made lights and switches into one minions file with shared mutex
